Allow re-ordering of windows within stack.
This affects where the windows are shown on screen (with our simple
tiling layout, at least!). This means that focus might jump to whichever
window happens to get placed under the cursor - not really what we want!
Disable tracking of the mouse for focus events while re-drawing a
layout to avoid this.