Define layouts/groups at the top level.
Introduce a new GroupBuilder to allow the real Group to be built after
we have an X Connection. Use some trickery to make a list of Box<Layout>
cloneable, as we need to give each group its own stack of layouts. Do as
QTile does, and reference the layouts by name - it means potential
errors at start-up, but I think that's unavoidable.