Clean-up logging.
Add some tracking for when we first/last see a window. Remove the
unhandled events log message as it's currently noise (- we can add it
back if it seems we're missing interesting events). Don't bother
tracking every configure_window request. We can add specific logging to
layouts if/when we need to debug them.